refactor(LottoNumbersInputView): 로또 번호를 입력받는 로직을 재사용하기 쉽게 분리
