+++ src/main/java/lotto/view/input/WinningNumberInputView.java
@@ -4,47 +4,21 @@ import lotto.number.LottoNumber;
 import lotto.number.LottoNumbers;
 import lotto.number.WinningNumbers;
 
-import java.util.Arrays;
-import java.util.List;
 import java.util.Scanner;
-import java.util.stream.Collectors;
 
 public class WinningNumberInputView {
 
 	private static final Scanner SCANNER = new Scanner(System.in);
+	private final LottoNumbersInputView lottoNumbersInputView;
+
+	public WinningNumberInputView() {
+		this.lottoNumbersInputView = LottoNumbersInputView.getInstance();
+	}
 
 	public WinningNumbers getWinningNumbers() {
-		LottoNumbers lottoNumbers = getLottoNumbers();
+		LottoNumbers lottoNumbers = lottoNumbersInputView.inputLottoNumbers("지난 주 당첨 번호를 입력해 주세요.");
 		LottoNumber bonusNumber = getBonusNumber();
 		return new WinningNumbers(lottoNumbers, bonusNumber);
-	}
-
-	private LottoNumbers getLottoNumbers() {
-		String input;
-		do {
-			System.out.println("지난 주 당첨 번호를 입력해 주세요.");
-			input = SCANNER.nextLine();
-		} while (!hasNoExceptionForLottoNumbers(input));
-		return createLottoNumbers(input);
-	}
-
-	private boolean hasNoExceptionForLottoNumbers(String input) {
-		try {
-			createLottoNumbers(input);
-			return true;
-		} catch (IllegalArgumentException e) {
-			System.out.println(e.getMessage());
-			return false;
-		}
-	}
-
-	private LottoNumbers createLottoNumbers(String input) {
-		input = input.replace(" ", "");
-		List<LottoNumber> lottoNumberList = Arrays.stream(input.split(","))
-				.map(Integer::parseInt)
-				.map(LottoNumber::new)
-				.collect(Collectors.toList());
-		return new LottoNumbers(lottoNumberList);
 	}
 
 	private LottoNumber getBonusNumber() {
